% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_js.R
\name{insert_js}
\alias{insert_js}
\title{Insert JavaScript}
\usage{
insert_js()
}
\description{
Insert the required JavaScript code from Server side
}
\details{
This function is a wrapper around shiny::insertUI() to add the JavaScript
from Server side. This way, no use_* call is needed in the UI.
}
\examples{
\dontrun{
insert_js()
}
}
