% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ktag_listener.R
\name{ktag_listener}
\alias{ktag_listener}
\title{Observe Event on Tag}
\usage{
ktag_listener(session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{session}{the session object from the server}
}
\value{
an observer reference class object (see shiny::observeEvent())
}
\description{
Observe Event on Tag
}
\examples{
\dontrun{
ktag_listener(session)
}
}
