% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{record}
\alias{record}
\title{Start Recording}
\usage{
record(input, session)
}
\arguments{
\item{input}{the input object of the Shiny server function}

\item{session}{the session object of the Shiny server function}
}
\value{
an observer reference class object (see shiny::observeEvent())
}
\description{
Gateway function to setup & init event recording in a single call.
}
\details{
\code{input} is required to create the listener that will capture events sent by
the JavaScript code.
\code{session} is required to register the onSessionEnded callback, get the URL
query string & use the token value as a session uuid.
}
\examples{
\dontrun{
ktag::record()
}
}
