% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_close.R
\name{register_close}
\alias{register_close}
\title{Log Session End}
\usage{
register_close(session)
}
\arguments{
\item{session}{the session object from the server}
}
\description{
Log Session End
}
\examples{
\dontrun{
register_close(session)
}
}
