% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_close.R
\name{session_close}
\alias{session_close}
\title{Log Session End}
\usage{
session_close(session)
}
\arguments{
\item{session}{the session object from the server}
}
\description{
Log Session End
}
\examples{
\dontrun{
session_close(session)
}
}
