% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_init.R
\name{session_init}
\alias{session_init}
\title{Log Session Start}
\usage{
session_init(session)
}
\arguments{
\item{session}{the session object from the server}
}
\description{
Log Session Start
}
\examples{
\dontrun{
session_init(session)
}
}
