% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ktag.R
\name{ktag}
\alias{ktag}
\title{Log User Actions}
\usage{
ktag(..., path = Sys.getenv("KTAG_PATH"))
}
\arguments{
\item{...}{parameters to log (see details)}

\item{path}{where to store the log file}
}
\description{
Log User Actions
}
\details{
The function understands the following parameters in ...
\itemize{
\item who: who is performing the action, session$token could be a pick
\item when: usually a timestamp, Sys.time() if omitted
\item where: from where the action is performed. When omitted, it will extract the namespace from what
\item what: the name of the action. Usually the id of the input, including its namespace
\item how: additional info to log, like the value of an input for example
}
}
\examples{
\dontrun{
ktag(who = session$token, what = "namespace-action")
}
}
