% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{record}
\alias{record}
\title{Start Recording}
\usage{
record(session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{session}{the session object of the Shiny server function}
}
\value{
an observer reference class object (see shiny::observeEvent())
}
\description{
Gateway function to setup & init event recording in a single call.
}
\details{
\code{session} is required to seteup the event listener, register the onSessionEnded callback,
get the URLcquery string & use the token value as a session uuid.
}
\examples{
\dontrun{
ktag::record()
}
}
